Auto-Attendence/pyproject.toml
[project]
name = "project55"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-generativeai>=0.8.4",
    "python-dotenv>=1.1.0",
    "streamlit>=1.44.1",
]
